requirements.txt
crawl4ai
pydanticwhere
